@@ -1,7 +1,6 @@
 package lru
 
 import (
-	"container/list"
 	"fmt"
 	"sync"
 
@@ -9,18 +8,18 @@ import (
 )
 
 // elementMap is an alias for a map from a generic interface to a list.Element.
-type elementMap map[interface{}]*list.Element
+type elementMap[K comparable, V any] map[K]V
 
 // entry represents a (key,value) pair entry in the Cache. The Cache's list
 // stores entries which let us get the cache key when an entry is evicted.
-type entry struct {
-	key   interface{}
-	value cache.Value
+type entry[K comparable, V cache.Value] struct {
+	key   K
+	value V
 }
 
 // Cache provides a generic thread-safe lru cache that can be used for
 // storing filters, blocks, etc.
-type Cache struct {
+type Cache[K comparable, V cache.Value] struct {
 	// capacity represents how much this cache can hold. It could be number
 	// of elements or a number of bytes, decided by the cache.Value's Size.
 	capacity uint64
@@ -30,11 +29,11 @@ type Cache struct {
 
 	// ll is a doubly linked list which keeps track of recency of used
 	// elements by moving them to the front.
-	ll *list.List
+	ll *List[entry[K, V]]
 
 	// cache is a generic cache which allows us to find an elements position
 	// in the ll list from a given key.
-	cache elementMap
+	cache elementMap[K, *Element[entry[K, V]]]
 
 	// mtx is used to make sure the Cache is thread-safe.
 	mtx sync.RWMutex
@@ -42,17 +41,17 @@ type Cache struct {
 
 // NewCache return a cache with specified capacity, the cache's size can't
 // exceed that given capacity.
-func NewCache(capacity uint64) *Cache {
-	return &Cache{
+func NewCache[K comparable, V cache.Value](capacity uint64) *Cache[K, V] {
+	return &Cache[K, V]{
 		capacity: capacity,
-		ll:       list.New(),
-		cache:    make(elementMap),
+		ll:       NewList[entry[K, V]](),
+		cache:    make(map[K]*Element[entry[K, V]]),
 	}
 }
 
 // evict will evict as many elements as necessary to make enough space for a new
 // element with size needed to be inserted.
-func (c *Cache) evict(needed uint64) (bool, error) {
+func (c *Cache[K, V]) evict(needed uint64) (bool, error) {
 	if needed > c.capacity {
 		return false, fmt.Errorf("can't evict %v elements in size, "+
 			"since capacity is %v", needed, c.capacity)
@@ -72,7 +71,7 @@ func (c *Cache) evict(needed uint64) (bool, error) {
 		// Find the least recently used item.
 		if elr := c.ll.Back(); elr != nil {
 			// Determine lru item's size.
-			ce := elr.Value.(*entry)
+			ce := elr.Value
 			es, err := ce.value.Size()
 			if err != nil {
 				return false, fmt.Errorf("couldn't determine "+
@@ -97,7 +96,7 @@ func (c *Cache) evict(needed uint64) (bool, error) {
 // exists, it will replace value and update it to be most recent item in cache.
 // The return value indicates whether items had to be evicted to make room for
 // the new element.
-func (c *Cache) Put(key interface{}, value cache.Value) (bool, error) {
+func (c *Cache[K, V]) Put(key K, value V) (bool, error) {
 	vs, err := value.Size()
 	if err != nil {
 		return false, fmt.Errorf("couldn't determine size of cache "+
@@ -115,7 +114,7 @@ func (c *Cache) Put(key interface{}, value cache.Value) (bool, error) {
 	// If the element already exists, remove it and decrease cache's size.
 	el, ok := c.cache[key]
 	if ok {
-		es, err := el.Value.(*entry).value.Size()
+		es, err := el.Value.value.Size()
 		if err != nil {
 			return false, fmt.Errorf("couldn't determine size of "+
 				"existing cache value %v", err)
@@ -132,7 +131,7 @@ func (c *Cache) Put(key interface{}, value cache.Value) (bool, error) {
 	}
 
 	// We have made enough space in the cache, so just insert it.
-	el = c.ll.PushFront(&entry{key, value})
+	el = c.ll.PushFront(entry[K, V]{key, value})
 	c.cache[key] = el
 	c.size += vs
 
@@ -141,25 +140,27 @@ func (c *Cache) Put(key interface{}, value cache.Value) (bool, error) {
 
 // Get will return value for a given key, making the element the most recently
 // accessed item in the process. Will return nil if the key isn't found.
-func (c *Cache) Get(key interface{}) (cache.Value, error) {
+func (c *Cache[K, V]) Get(key K) (V, error) {
 	c.mtx.Lock()
 	defer c.mtx.Unlock()
+
+	var defaultVal V
 
 	el, ok := c.cache[key]
 	if !ok {
 		// Element not found in the cache.
-		return nil, cache.ErrElementNotFound
+		return defaultVal, cache.ErrElementNotFound
 	}
 
 	// When the cache needs to evict a element to make space for another
 	// one, it starts eviction from the back, so by moving this element to
 	// the front, it's eviction is delayed because it's recently accessed.
 	c.ll.MoveToFront(el)
-	return el.Value.(*entry).value, nil
+	return el.Value.value, nil
 }
 
 // Len returns number of elements in the cache.
-func (c *Cache) Len() int {
+func (c *Cache[K, V]) Len() int {
 	c.mtx.RLock()
 	defer c.mtx.RUnlock()
 
