@@ -1,4 +1,4 @@
-package cache_test
+package neutrino
 
 import (
 	"crypto/rand"
@@ -28,8 +28,12 @@ func TestBlockFilterCaches(t *testing.T) {
 
 	// Initialize all types of caches we want to test, for both filters and
 	// blocks. Currently the LRU cache is the only implementation.
-	filterCaches := []cache.Cache{lru.NewCache(cacheSize)}
-	blockCaches := []cache.Cache{lru.NewCache(cacheSize)}
+	filterCaches := []cache.Cache[FilterCacheKey, *CacheableFilter]{
+		lru.NewCache[FilterCacheKey, *CacheableFilter](cacheSize),
+	}
+	blockCaches := []cache.Cache[wire.InvVect, *CacheableBlock]{
+		lru.NewCache[wire.InvVect, *CacheableBlock](cacheSize),
+	}
 
 	// Generate a list of hashes, filters and blocks that we will use as
 	// cache keys an values.
@@ -55,9 +59,9 @@ func TestBlockFilterCaches(t *testing.T) {
 		filters = append(filters, filter)
 
 		// Put the generated filter in the filter caches.
-		cacheKey := cache.FilterCacheKey{blockHash, filterType}
+		cacheKey := FilterCacheKey{blockHash, filterType}
 		for _, c := range filterCaches {
-			_, _ = c.Put(cacheKey, &cache.CacheableFilter{Filter: filter})
+			_, _ = c.Put(cacheKey, &CacheableFilter{Filter: filter})
 		}
 
 		msgBlock := &wire.MsgBlock{}
@@ -70,7 +74,7 @@ func TestBlockFilterCaches(t *testing.T) {
 			wire.InvTypeWitnessBlock, &blockHash,
 		)
 		for _, c := range blockCaches {
-			_, _ = c.Put(*blockKey, &cache.CacheableBlock{block})
+			_, _ = c.Put(*blockKey, &CacheableBlock{block})
 		}
 	}
 
@@ -80,7 +84,7 @@ func TestBlockFilterCaches(t *testing.T) {
 		blockHash := blockHash
 
 		// Check filter caches.
-		cacheKey := cache.FilterCacheKey{blockHash, filterType}
+		cacheKey := FilterCacheKey{blockHash, filterType}
 		for _, c := range filterCaches {
 			e, err := c.Get(cacheKey)
 			if err != nil {
@@ -88,7 +92,7 @@ func TestBlockFilterCaches(t *testing.T) {
 			}
 
 			// Ensure we got the correct filter.
-			filter := e.(*cache.CacheableFilter).Filter
+			filter := e.Filter
 			if filter != filters[i] {
 				t.Fatalf("Filters not equal: %v vs %v ",
 					filter, filters[i])
@@ -106,7 +110,7 @@ func TestBlockFilterCaches(t *testing.T) {
 			}
 
 			// Ensure it is the same block.
-			block := b.(*cache.CacheableBlock).Block
+			block := b.Block
 			if block != blocks[i] {
 				t.Fatalf("Not equal: %v vs %v ",
 					block, blocks[i])
